Improve search for open-circuit initial conditions
This update aims to improve the way in which initial conditions for simulations starting from open-circuit are computed, particularly for simulations that begin in the dark. The main difference is that find_Voc searches for the Voc by increasing the voltage from 0V, rather than decreasing the voltage from Vbi. A second attempt is then made to refine the solution subject to a stricter AbsTol, which seems to help when DI=0.